1) Implemented BookingTargetsInfo handling 2) added constants for IXSI providers etc

+++ src/main/java/de/rwth/idsg/bikeman/ixsi/dto/query/StationDTO.java
@@ -1,9 +1,11 @@
 package de.rwth.idsg.bikeman.ixsi.dto.query;
 
+import de.rwth.idsg.bikeman.ixsi.schema.NonNegativeInteger;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 /**
  * Created by max on 01/10/14.
@@ -14,7 +16,7 @@ public class StationDTO {
     private String stationId;
     private BigDecimal location_longitude;
     private BigDecimal location_latitude;
-    private int slotCount;
+    private BigInteger slotCount;
     private String name;
     private String providerId;
     private String description;
